requirements.txt
Flask==2.3.3
google-cloud-dialogflow-cx==1.25.0
requests==2.31.0
gunicorn==21.2.0
